requirements.txt
numpy==1.21.3
Pillow==9.2.0
streamlit==1.12.2
tensorflow==2.10.0
